requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-huggingface
sentence-transformers
faiss-cpu
pypdf
python-dotenv